requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Werkzeug==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.0.2
Python-DotEnv==1.0.1
Flask-Mail==0.9.1
Celery==5.3
Redis==5.0.1